Only merge with GET parameters
When using this module on a page that has GET and POST variables. All variables GET and POST variables will be added to the query string. This fix will only add the GET variables. Values from the Request::query() array will be overwritten by $parameters array, so sort and field values will be updated correctly.